Erlang R16B03 -> 20.3: replace crypto:rand_bytes/1
crypto:rand_bytes/1 is removed in 20th erlang so it should be
replaced with crypto:strong_rand_bytes/1. The problem is that
crypto:strong_rand_bytes/1 can throw a low_entropy exception.
In this case we try to increase entropy using rand_seed function.

It's important that crypto:rand_seed/1 doesn't set the state of
random generator (as it was for random:seed/1 for example) but
only adds entropy to current state. It means that it will not
get worse if the generator is seeded with not very random stuff.

Change-Id: I1c7398e3ee9b1b4de1119c7265935cb5035de14b