Erlang R16B03 -> 20.3: replace crypto:rand_bytes/1
crypto:rand_bytes/1 is removed in 20th erlang so it should be
replaced with crypto:strong_rand_bytes/1. The problem is that
crypto:strong_rand_bytes/1 can throw a low_entropy exception.
In this case we try to increase enthropy using rand_seed function

The utility function can be found by the following link:
http://review.couchbase.org/#/c/92351/

Change-Id: I1c7398e3ee9b1b4de1119c7265935cb5035de14b