Memcached auth server
In order to handle auth requests from memcached we need to start
a server that establishes a connection to memcached (with
'RbacProvoder' feature enabled) and waits for auth commands from it.
At this point the commands could be the following:
1) AuthRequest - general authentication command
2) GetRbacPermissions - return rbac record for specified user
(implementation of the commands is not included in the commit)
We can run several instancies of that process if we need to scale.

Change-Id: Id8628b82c51c111a3ef892b50e2bd90d3ea133dd