GetUserPermission command implementation
Get the RBAC entry for the specified user. The user is specified as:

 {"username":"domain"}

Domain may be "local" or "external"

The request contains the name of the requested user, and the response
should contain the username in the key field, and the JSON containing
the RBAC definition for the user in the body.

Change-Id: I525cdbf4a9f2bf56a8a508dd8c0a46822b1c60b7